Rearrange data in test files
